safari-600.7-branch/LayoutTests/platform/ios-sim/http/tests/media/video-play-stall-before-meta-data-expected.txt

Test that stalling very early, while loading meta-data, stops delaying the load event.

EVENT(beforeload)
EVENT(loadstart)
EVENT(window.onload)
END OF TEST
